backend/requirements.txt
asgiref==3.5.2
Babel==2.10.3
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.2.0
django-money==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
enmerkar==0.7.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.2.1
packaging==21.3
Pillow==9.2.0
prices==1.1.0
psycopg2==2.9.3
py-moneyed==2.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.8.1
